Update to conform to new changes in standard
