+++ Nivel1_P1.py
@@ -8,42 +8,6 @@ import random
 from config import *
 from entities import *
 
-class bloques:
-    def __init__(self, Dibujo):
-        self.Dibujo = Dibujo
-        self.bloque = pygame.image.load("piedra_2.jpg").convert()
-        self.bloque2 = pygame.image.load("piedra_2.jpg").convert()
-        self.bloque3 = pygame.image.load("piedra_2.jpg").convert()
-        self.bloque4 = pygame.image.load("piedra_2.jpg").convert()
-
-        self.x = 400
-        self.y = 40
-
-        self.bloque5 = pygame.image.load("piedra_2.jpg").convert()
-        self.bloque6 = pygame.image.load("piedra_2.jpg").convert()
-        self.bloque7 = pygame.image.load("piedra_2.jpg").convert()
-
-        self.bloque8 = pygame.image.load("piedra_2.jpg").convert()
-        self.bloque9 = pygame.image.load("piedra_2.jpg").convert()
-
-
-    def dibujar_pared(self):
-        self.Dibujo.blit(self.bloque, (self.x, self.y))
-        self.Dibujo.blit(self.bloque2, (self.x, self.y+40))
-        self.Dibujo.blit(self.bloque3, (self.x, self.y+80))
-        self.Dibujo.blit(self.bloque4, (self.x, self.y+120))
-
-
-        self.Dibujo.blit(self.bloque5, (self.x+40, self.y+120))
-        self.Dibujo.blit(self.bloque6, (self.x+80, self.y+120))
-        self.Dibujo.blit(self.bloque7, (self.x+120, self.y+120))
-
-        self.Dibujo.blit(self.bloque9, (self.x+380, self.y+300))
-        self.Dibujo.blit(self.bloque8, (self.x+420, self.y+300))
-        
-        pygame.display.flip()
-
-
 class Game:
     def __init__(self):
         pygame.init()
@@ -51,14 +15,6 @@ class Game:
 
         pygame.mixer.init()
         self.musica_fondo()
-
-        self.surface = pygame.display.set_mode((1000, 520))
-        self.snake = Snake(self.surface)
-        self.snake.dibujar_snake()
-        self.manzana = Manzana(self.surface)
-        self.manzana.dibujar_manzana()
-        self.bloque = bloques(self.surface)
-        self.bloque.dibujar_pared()
 
     def musica_fondo(self):
         pygame.mixer.music.load("Recursos/bg_music_1.mp3")
@@ -74,7 +30,6 @@ class Game:
     def reset(self):
         self.snake = Snake(self.surface)
         self.manzana = Manzana(self.surface)
-        self.bloque = bloques(self.surface)
 
 
     def colision(self, x1, y1, x2, y2):
@@ -91,14 +46,13 @@ class Game:
         self.fondo_juego()
         self.snake.mover_snake()
         self.manzana.dibujar_manzana()
-        self.bloque.dibujar_pared()
         self.score()
         pygame.display.flip()
 
         # Comer Manzana
         for i in range(self.snake.largo):
             if self.colision(self.snake.x[i], self.snake.y[i], self.manzana.x, self.manzana.y):
-                self.musica_juego("ding")
+                self.musica_juego("di   ng")
                 self.snake.incrementar_largo()
                 self.manzana.mover_manzana()
 
@@ -109,10 +63,9 @@ class Game:
                 raise "Se toca"
 
         # Toca pared
-        if not (0 <= self.snake.x[0] <= 1000 and 0 <= self.snake.y[0] <= 520):
+        if not (0 <= self.snake.x[0] <= 1000 and 0 <= self.snake.y[0] <= 500):
             self.musica_juego('crash')
             raise "Toco Pared"
-
 
     def score(self):
         font = pygame.font.SysFont('arial',30)
@@ -129,9 +82,13 @@ class Game:
         pygame.mixer.music.pause()
         pygame.display.flip()
 
-    
-
     def run(self):
+        self.surface = pygame.display.set_mode((1000, 500))
+        self.snake = Snake(self.surface)
+        self.snake.dibujar_snake()
+        self.manzana = Manzana(self.surface)
+        self.manzana.dibujar_manzana()
+        
         running = True
         pausa = False
 
@@ -158,16 +115,6 @@ class Game:
                         if event.key == K_DOWN:
                             self.snake.mover_abajo()
 
-                if self.snake.x[0] == 400 and self.snake.y[0] == 40:
-                    running = False
-                elif self.snake.x[0] == 400 and self.snake.y[0] == 80:
-                    running = False
-                elif self.snake.x[0] == 400 and self.snake.y[0] == 120:
-                    running = False
-                elif self.snake.x[0] == 400 and self.snake.y[0] == 160:
-                    running = False
-
-            
                 elif event.type == QUIT:
                     running = False
             try:
@@ -180,9 +127,6 @@ class Game:
                 pausa = True
                 self.reset()
             time.sleep(.1)
-
-
-
 
 if __name__ == '__main__':
     game = Game()
