+++ Nivel2_P1.py
@@ -1,3 +1,4 @@
+#Intento 3
 
 import pygame
 from pygame.locals import *
@@ -8,7 +9,6 @@ from config import *
 from entities import *
 
 class bloques:
-
     def __init__(self, Dibujo):
         self.Dibujo = Dibujo
         self.bloque = pygame.image.load("piedra_2.jpg").convert()
@@ -16,8 +16,8 @@ class bloques:
         self.bloque3 = pygame.image.load("piedra_2.jpg").convert()
         self.bloque4 = pygame.image.load("piedra_2.jpg").convert()
 
-        self.x = 200
-        self.y = 200
+        self.x = 400
+        self.y = 40
 
         self.bloque5 = pygame.image.load("piedra_2.jpg").convert()
         self.bloque6 = pygame.image.load("piedra_2.jpg").convert()
@@ -26,43 +26,21 @@ class bloques:
         self.bloque8 = pygame.image.load("piedra_2.jpg").convert()
         self.bloque9 = pygame.image.load("piedra_2.jpg").convert()
 
-        self.bloque10 = pygame.image.load("piedra_2.jpg").convert()
-        self.bloque11 = pygame.image.load("piedra_2.jpg").convert()
-        self.bloque12= pygame.image.load("piedra_2.jpg").convert()
-        self.bloque13 = pygame.image.load("piedra_2.jpg").convert()
-
-        self.bloque14 = pygame.image.load("piedra_2.jpg").convert()
-        self.bloque15 = pygame.image.load("piedra_2.jpg").convert()
-        self.bloque16= pygame.image.load("piedra_2.jpg").convert()
-        self.bloque17 = pygame.image.load("piedra_2.jpg").convert()
-
-
 
     def dibujar_pared(self):
         self.Dibujo.blit(self.bloque, (self.x, self.y))
-        self.Dibujo.blit(self.bloque2, (self.x+40, self.y+40))
-        self.Dibujo.blit(self.bloque3, (self.x+80, self.y+80))
-        self.Dibujo.blit(self.bloque4, (self.x+120, self.y+120))
+        self.Dibujo.blit(self.bloque2, (self.x, self.y+40))
+        self.Dibujo.blit(self.bloque3, (self.x, self.y+80))
+        self.Dibujo.blit(self.bloque4, (self.x, self.y+120))
 
 
-        self.Dibujo.blit(self.bloque5, (self.x+200, self.y+120))
-        self.Dibujo.blit(self.bloque6, (self.x+240, self.y+160))
-        self.Dibujo.blit(self.bloque7, (self.x+280, self.y+200))
+        self.Dibujo.blit(self.bloque5, (self.x+40, self.y+120))
+        self.Dibujo.blit(self.bloque6, (self.x+80, self.y+120))
+        self.Dibujo.blit(self.bloque7, (self.x+120, self.y+120))
 
-        self.Dibujo.blit(self.bloque9, (self.x+380, self.y+280))
-        self.Dibujo.blit(self.bloque8, (self.x+420, self.y+280))
-
-        self.Dibujo.blit(self.bloque10, (self.x+400, self.y-160))
-        self.Dibujo.blit(self.bloque11, (self.x+440, self.y-160))
-        self.Dibujo.blit(self.bloque12, (self.x+480, self.y-160))
-        self.Dibujo.blit(self.bloque13, (self.x+520, self.y-160))
-
-        self.Dibujo.blit(self.bloque14, (self.x+560, self.y-160))
-        self.Dibujo.blit(self.bloque15, (self.x+560, self.y-140))
-        self.Dibujo.blit(self.bloque16, (self.x+560, self.y-120))
-        self.Dibujo.blit(self.bloque17, (self.x+560, self.y-100))
-
-
+        self.Dibujo.blit(self.bloque9, (self.x+380, self.y+300))
+        self.Dibujo.blit(self.bloque8, (self.x+420, self.y+300))
+        
         pygame.display.flip()
 
 
@@ -123,11 +101,10 @@ class Game:
                 raise "Se toca"
 
         # Toca pared
-        if not (0 <= self.snake.x[0] <= 960 and 0 <= self.snake.y[0] <= 500):
+        if not (0 <= self.snake.x[0] <= 1000 and 0 <= self.snake.y[0] <= 520):
             self.musica_juego('crash')
             raise "Toco Pared"
 
-        print(self.snake.y[0])
 
     def score(self):
         font = pygame.font.SysFont('arial',30)
@@ -181,6 +158,16 @@ class Game:
                         if event.key == K_DOWN:
                             self.snake.mover_abajo()
 
+                if self.snake.x[0] == 400 and self.snake.y[0] == 40:
+                    running = False
+                elif self.snake.x[0] == 400 and self.snake.y[0] == 80:
+                    running = False
+                elif self.snake.x[0] == 400 and self.snake.y[0] == 120:
+                    running = False
+                elif self.snake.x[0] == 400 and self.snake.y[0] == 160:
+                    running = False
+
+            
                 elif event.type == QUIT:
                     running = False
             try:
@@ -193,8 +180,6 @@ class Game:
                 pausa = True
                 self.reset()
             time.sleep(.1)
-
-
 
 
 if __name__ == '__main__':
