+++ test.py
@@ -1,0 +1,14 @@
+from Nivel1_P1 import Game as n1
+from Nivel2_P1 import Game as n2
+from Nivel3_P1 import Game as n3
+from Nivel4_P1 import Game as n4
+
+if __name__ == '__main__':
+    game = n1()
+    game.run()
+    game = n2()
+    game.run()
+    game = n3()
+    game.run()
+    game = n4()
+    game.run()
